Rename files to kebab-case for consistency with the rest of repo
