+++ packages/webpack-plugin/src/get-module-in-graph.ts
