+++ packages/webpack-plugin/src/plugin-options.ts
