+++ packages/webpack-plugin/src/stylable-bootstrap-module.ts
@@ -2,8 +2,8 @@ import { EOL } from 'os';
 import webpack from 'webpack';
 import { RawSource } from 'webpack-sources';
 import { WEBPACK_STYLABLE } from './runtime-dependencies';
+import { StylableImportDependency } from './stylable-dependencies';
 import { getStylableModulesFromDependencies, renderStaticCSS } from './stylable-module-helpers';
-import { StylableImportDependency } from './StylableDependencies';
 import { StylableModule, StylableWebpackPluginOptions } from './types';
 const Module = require('webpack/lib/Module');
