+++ packages/webpack-plugin/src/stylable-dependencies.ts
