+++ packages/webpack-plugin/src/stylable-generator.ts
