+++ packages/webpack-plugin/src/stylable-webpack-plugin.ts
@@ -4,15 +4,15 @@ import { StylableOptimizer } from '@stylable/optimizer';
 import { EOL } from 'os';
 import webpack from 'webpack';
 import { RawSource } from 'webpack-sources';
-import { getModuleInGraph, hasStylableModuleInGraph } from './getModuleInGraph';
-import { normalizeOptions } from './PluginOptions';
+import { getModuleInGraph, hasStylableModuleInGraph } from './get-module-in-graph';
+import { normalizeOptions } from './plugin-options';
 import { RUNTIME_SOURCE, WEBPACK_STYLABLE } from './runtime-dependencies';
+import { StylableBootstrapModule } from './stylable-bootstrap-module';
+import { StylableAssetDependency, StylableImportDependency } from './stylable-dependencies';
+import { StylableGenerator } from './stylable-generator';
 import { calculateModuleDepthAndShallowStylableDependencies, renderStaticCSS } from './stylable-module-helpers';
-import { StylableBootstrapModule } from './StylableBootstrapModule';
-import { StylableAssetDependency, StylableImportDependency } from './StylableDependencies';
-import { StylableGenerator } from './StylableGenerator';
-import { StylableParser } from './StylableParser';
-import { StyleableAutoInitDependency, StyleableAutoInitDependencyTemplate } from './StyleableAutoInitDependency';
+import { StylableParser } from './stylable-parser';
+import { StyleableAutoInitDependency, StyleableAutoInitDependencyTemplate } from './styleable-auto-init-dependency';
 import { CalcResult, ShallowPartial, StylableModule, StylableWebpackPluginOptions } from './types';
 import { isImportedByNonStylable } from './utils';
