+++ packages/webpack-plugin/src/styleable-auto-init-dependency.ts
